third_party/WebKit/LayoutTests/external/wpt/web-animations/interfaces/AnimationPlaybackEvent/constructor-expected.txt

This is a testharness.js-based test.
FAIL Event created without an event parameter has null time values assert_equals: expected (object) null but got (number) 0
PASS Created event reflects times specified in constructor
Harness: the test ran to completion.
